+++ data-raw/StdIntLit.R
@@ -1,0 +1,3 @@
+StdIntLit <- read.csv("data-raw/StdIntLit.csv",stringsAsFactors=FALSE)
+str(StdIntLit)
+save(StdIntLit,file="data/StdIntLit.RData")