+++ tests/testthat/test_OUTPUTS.R
@@ -270,28 +270,42 @@ test_that("backCalc() output types",{
 
 test_that("backCalc() output values",{
   ## First fish, first increment
-  ### Dahl-Lea
   tmp <- SMBassWB[1,]
+  ### Dahl-Lea
   capture.output(out <- backCalc(tmp,lencap,BCM="DALE",inFormat="wide",
                                  outFormat="long",digits=1))
   exp1 <- round(with(tmp,lencap*rad1/radcap),1)
   expect_equal(out$bclen[1],exp1)
-  ### Fraser-Lea
+  ### Fraser-Lee
+  a <- aStandard("Smallmouth Bass")
   tmp <- SMBassWB[1,]
-  capture.output(out <- backCalc(tmp,lencap,BCM="FRALE",a=15,inFormat="wide",
+  capture.output(out <- backCalc(tmp,lencap,BCM="FRALE",a=a,inFormat="wide",
                                  outFormat="long",digits=1))
-  exp1 <- round(with(tmp,(lencap-15)*rad1/radcap+15),1)
+  exp1 <- round(with(tmp,(lencap-a)*rad1/radcap+a),1)
   expect_equal(out$bclen[1],exp1)
   
   ## Last fish, first and ninth increment
-  ### Dahl-Lea
   tmp <- SMBassWB[181,]
-  capture.output(out <- backCalc(tmp,lencap,BCM="FRALE",a=15,inFormat="wide",
+  ### Dahl-Lea
+  capture.output(out <- backCalc(tmp,lencap,BCM="DALE",inFormat="wide",
                                  outFormat="long",digits=1))
-  exp1 <- round(with(tmp,(lencap-15)*rad1/radcap+15),1)
+  exp1 <- round(with(tmp,lencap*rad1/radcap),1)
   expect_equal(out$bclen[1],exp1)
-  exp9 <- round(with(tmp,(lencap-15)*rad9/radcap+15),1)
+  exp9 <- round(with(tmp,lencap*rad9/radcap),1)
   expect_equal(out$bclen[9],exp9)
+  ### Fraser-Lee
+  capture.output(out <- backCalc(tmp,lencap,BCM="FRALE",a=a,inFormat="wide",
+                                 outFormat="long",digits=1))
+  exp1 <- round(with(tmp,(lencap-a)*rad1/radcap+a),1)
+  expect_equal(out$bclen[1],exp1)
+  exp9 <- round(with(tmp,(lencap-a)*rad9/radcap+a),1)
+  expect_equal(out$bclen[9],exp9)
+})
+
+
+test_that("aStandard() outputs",{
+  expect_type(aStandard("Bluegill"),"integer")
+  expect_equal(aStandard("Bluegill"),20)
 })